email-outreach/requirements.txt
sendgrid>=6.11.0
gspread>=5.12.0
google-auth>=2.23.0
openpyxl>=3.1.2
pandas>=2.1.0
